+++ .eslintrc.cjs
@@ -1,9 +1,0 @@
-/** @type {import('@types/eslint').Linter.BaseConfig} */
-module.exports = {
-	extends: [
-		require('prettier'),
-	],
-	env: {
-		node: true,
-	}
-};