+++ src/IndexNameManager/IndexNameManager.ts
@@ -1,3 +1,6 @@
+export type IndexName = string;
+export type ColumnName = string;
+
 export type IndexNameAttributes = {
   name: IndexName;
   separator?: string;
@@ -112,44 +115,3 @@ export default class IndexNameManager {
     });
   }
 }
-
-type LowercaseLetter =
-  | 'a'
-  | 'b'
-  | 'c'
-  | 'd'
-  | 'e'
-  | 'f'
-  | 'g'
-  | 'h'
-  | 'i'
-  | 'j'
-  | 'k'
-  | 'l'
-  | 'm'
-  | 'n'
-  | 'o'
-  | 'p'
-  | 'q'
-  | 'r'
-  | 's'
-  | 't'
-  | 'u'
-  | 'v'
-  | 'w'
-  | 'x'
-  | 'y'
-  | 'z';
-
-type UppercaseLetter = Uppercase<LowercaseLetter>;
-
-// Matches only: a–z, A–Z, 0–9, underscore (_), dash (-)
-type AllowedColumnChars = LowercaseLetter | UppercaseLetter | Digit | '_' | '-';
-
-// Matches only: a–z, 0–9, underscore (_), dash (-)
-type AllowedIndexChars = LowercaseLetter | Digit | '_' | '-';
-
-type Digit = '0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9';
-
-export type IndexName = `${AllowedIndexChars}${string}`;
-export type ColumnName = `${AllowedColumnChars}${string}`;