+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Change Log
 
-### v1.0.0 on 20XX-XX-XX
+### v2.0.0-beta.1 on 2025-08-14
 
-- Initial release
+- Initial release with IndexManager and nested field support
+
+### v1.0.0-beta.1 on 2022
+
+- Test release of Query Builder